When switching to a model in rspec-mode stay in rspec-mode
I don't have rspec-mode on by default because I don't always use rspec.
Instead I activate it when in a spec file. If I have rspec-mode activated
and I jump to a model through it I want to be in rspec-mode when looking at
that model also (so I can easily jump back)